\name{Rcpp Modules Examples}
\alias{Histogram2D}
\alias{Solver}
\alias{distanceDF}
\alias{Rcpp_Solver-class}
\alias{Rcpp_Histogram2D-class}
\title{
  Functions and Objects created by Rcpp Modules Example
}
\description{
  These function and objects are accessible from R via the Rcpp Modules mechanism
  which creates them based on the declaration in the C++ file.  
}
\seealso{
  The Rcpp Modules vignette.
}